Fix uint8 overflow in SAE present_elements detection
F.sum(0) was computed in uint8 arithmetic, which overflows when total
atom counts exceed 255. This could silently drop elements from the
output SAE file. Cast to int64 before summing.

+++ aimnet2mult/data/sgdataset.py
@@ -388,7 +388,7 @@ class SizeGroupedDataset:
             F = np.concatenate([eye[g[numbers_key]].sum(-2)
                                 for g in self.values()])
             sap = np.linalg.lstsq(F, E, rcond=None)[0]
-            present_elements = np.nonzero(F.sum(0))[0]
+            present_elements = np.nonzero(F.astype(np.int64).sum(0))[0]
         else:
             ntyp = max(sap_dict.keys()) + 1
             sap = np.zeros(ntyp) * np.nan
